// generated-by-analysis/Pilot10/graph-inputs/scenario-recency-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/scenario-recency.ts.
// spell-checker: disable

export const labels = [
	"Breached Device",
	"Locked Device",
	"Breached Account",
	"Locked Account",
	"Breached Social",
	"Locked Social",
	"Breached Financial",
	"Locked Financial",
	"Breached Passwords",
	"Locked Passwords",
	"Replaced Device/OS",
	"Broken Promise",
	"Abuse"
] as const; 
export const absoluteData = {
	"Within Year": [
		0,
		1,
		0,
		0,
		0,
		1,
		0,
		0,
		0,
		0,
		1,
		2,
		1
	],
	"1-3 Years": [
		1,
		0,
		2,
		1,
		1,
		1,
		2,
		2,
		1,
		1,
		1,
		1,
		0
	],
	"3-5 Years": [
		0,
		1,
		0,
		1,
		1,
		0,
		0,
		1,
		1,
		1,
		1,
		2,
		0
	],
	"5-10 Years": [
		1,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		1,
		1,
		1
	],
	"> 10 years": [
		1,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0
	],
	"Other": [
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		1
	]
} as const;
export const percentData = {
	"Within Year": [
		0,
		50,
		0,
		0,
		0,
		50,
		0,
		0,
		0,
		0,
		25,
		33.333333333333336,
		33.333333333333336
	],
	"1-3 Years": [
		33.333333333333336,
		0,
		100,
		50,
		50,
		50,
		100,
		66.66666666666667,
		50,
		50,
		25,
		16.666666666666668,
		0
	],
	"3-5 Years": [
		0,
		50,
		0,
		50,
		50,
		0,
		0,
		33.333333333333336,
		50,
		50,
		25,
		33.333333333333336,
		0
	],
	"5-10 Years": [
		33.333333333333336,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		25,
		16.666666666666668,
		33.333333333333336
	],
	"> 10 years": [
		33.333333333333336,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0
	],
	"Other": [
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		33.333333333333336
	]
} as const;