// generated-by-analysis/Pilot10/graph-inputs/recovery-duration-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/recovery-duration.ts.
// spell-checker: disable

export const labels = [
	"Breached Device",
	"Locked Device",
	"Breached Account",
	"Locked Account",
	"Breached Social",
	"Locked Social",
	"Breached Financial",
	"Locked Financial",
	"Breached Passwords",
	"Locked Passwords",
	"Replaced Device/OS",
	"Broken Promise",
	"Abuse"
] as const;
export const absoluteData = {
	"Never Regained": [
		4,
		2,
		0,
		1,
		1,
		3,
		1,
		1,
		1,
		0,
		6,
		6,
		2
	],
	"Eventually": [
		1,
		2,
		0,
		2,
		2,
		1,
		1,
		0,
		2,
		5,
		3,
		2,
		3
	],
	"Within a Week": [
		1,
		2,
		3,
		2,
		1,
		2,
		1,
		4,
		1,
		2,
		2,
		4,
		1
	],
	"Within a Day": [
		2,
		3,
		1,
		1,
		1,
		1,
		1,
		2,
		1,
		1,
		0,
		7,
		0
	],
	"Other": [
		0,
		0,
		0,
		1,
		1,
		0,
		1,
		0,
		1,
		0,
		0,
		0,
		1
	],
	"Never Lost": [
		0,
		0,
		0,
		0,
		0,
		0,
		1,
		0,
		0,
		0,
		0,
		0,
		0
	]
} as const;
export const percentData = {
	"Never Regained": [
		50,
		22.22222222222222,
		0,
		14.285714285714286,
		16.666666666666668,
		42.857142857142854,
		16.666666666666668,
		14.285714285714286,
		16.666666666666668,
		0,
		54.54545454545455,
		31.57894736842105,
		28.571428571428573
	],
	"Eventually": [
		12.5,
		22.22222222222222,
		0,
		28.571428571428573,
		33.333333333333336,
		14.285714285714286,
		16.666666666666668,
		0,
		33.333333333333336,
		62.5,
		27.272727272727273,
		10.526315789473685,
		42.857142857142854
	],
	"Within a Week": [
		12.5,
		22.22222222222222,
		75,
		28.571428571428573,
		16.666666666666668,
		28.571428571428573,
		16.666666666666668,
		57.142857142857146,
		16.666666666666668,
		25,
		18.181818181818183,
		21.05263157894737,
		14.285714285714286
	],
	"Within a Day": [
		25,
		33.333333333333336,
		25,
		14.285714285714286,
		16.666666666666668,
		14.285714285714286,
		16.666666666666668,
		28.571428571428573,
		16.666666666666668,
		12.5,
		0,
		36.8421052631579,
		0
	],
	"Other": [
		0,
		0,
		0,
		14.285714285714286,
		16.666666666666668,
		0,
		16.666666666666668,
		0,
		16.666666666666668,
		0,
		0,
		0,
		14.285714285714286
	],
	"Never Lost": [
		0,
		0,
		0,
		0,
		0,
		0,
		16.666666666666668,
		0,
		0,
		0,
		0,
		0,
		0
	]
} as const;