// generated-by-analysis/Pilot10/graph-inputs/likert-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/likert.ts.
// spell-checker: disable

	export const labels = [
	"Breached Device",
	"Locked Device",
	"Breached Account",
	"Locked Account",
	"Breached Social",
	"Locked Social",
	"Breached Financial",
	"Locked Financial",
	"Breached Passwords",
	"Locked Passwords",
	"Replaced Device/OS",
	"Broken Promise",
	"Abuse",
	"Lost Photos",
	"Lost Emails",
	"Mental Health"
] as const;
	export const counts = {
	"1": [
		0,
		2,
		0,
		1,
		0,
		2,
		0,
		1,
		1,
		0,
		2,
		2,
		1,
		1,
		2,
		0
	],
	"2": [
		0,
		0,
		1,
		1,
		0,
		1,
		1,
		1,
		0,
		1,
		2,
		4,
		0,
		1,
		0,
		0
	],
	"3": [
		2,
		1,
		1,
		1,
		1,
		1,
		0,
		1,
		1,
		0,
		3,
		5,
		0,
		0,
		2,
		2
	],
	"4": [
		3,
		2,
		0,
		0,
		1,
		0,
		0,
		1,
		1,
		2,
		0,
		3,
		2,
		0,
		1,
		1
	],
	"5": [
		0,
		1,
		1,
		2,
		2,
		3,
		3,
		0,
		0,
		2,
		1,
		0,
		2,
		3,
		1,
		7
	],
	"6": [
		1,
		1,
		1,
		2,
		2,
		0,
		0,
		2,
		2,
		2,
		3,
		4,
		2,
		1,
		0,
		1
	],
	"7": [
		2,
		2,
		0,
		0,
		0,
		0,
		2,
		1,
		1,
		1,
		0,
		1,
		0,
		1,
		1,
		0
	],
	"No loss": [
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0
	]
} as const;
	export const percentsOfResponses = {
	"1": [
		0,
		8,
		0,
		4,
		0,
		8,
		0,
		4,
		4,
		0,
		8,
		8,
		4,
		4,
		8,
		0
	],
	"2": [
		0,
		0,
		4,
		4,
		0,
		4,
		4,
		4,
		0,
		4,
		8,
		16,
		0,
		4,
		0,
		0
	],
	"3": [
		8,
		4,
		4,
		4,
		4,
		4,
		0,
		4,
		4,
		0,
		12,
		20,
		0,
		0,
		8,
		8
	],
	"4": [
		12,
		8,
		0,
		0,
		4,
		0,
		0,
		4,
		4,
		8,
		0,
		12,
		8,
		0,
		4,
		4
	],
	"5": [
		0,
		4,
		4,
		8,
		8,
		12,
		12,
		0,
		0,
		8,
		4,
		0,
		8,
		12,
		4,
		28
	],
	"6": [
		4,
		4,
		4,
		8,
		8,
		0,
		0,
		8,
		8,
		8,
		12,
		16,
		8,
		4,
		0,
		4
	],
	"7": [
		8,
		8,
		0,
		0,
		0,
		0,
		8,
		4,
		4,
		4,
		0,
		4,
		0,
		4,
		4,
		0
	],
	"No loss": [
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0
	]
} as const;
	export const percentsOfAnswered = {
	"1": [
		0,
		22.22222222222222,
		0,
		14.285714285714286,
		0,
		28.571428571428573,
		0,
		14.285714285714286,
		16.666666666666668,
		0,
		18.181818181818183,
		10.526315789473685,
		14.285714285714286,
		14.285714285714286,
		28.571428571428573,
		0
	],
	"2": [
		0,
		0,
		25,
		14.285714285714286,
		0,
		14.285714285714286,
		16.666666666666668,
		14.285714285714286,
		0,
		12.5,
		18.181818181818183,
		21.05263157894737,
		0,
		14.285714285714286,
		0,
		0
	],
	"3": [
		25,
		11.11111111111111,
		25,
		14.285714285714286,
		16.666666666666668,
		14.285714285714286,
		0,
		14.285714285714286,
		16.666666666666668,
		0,
		27.272727272727273,
		26.31578947368421,
		0,
		0,
		28.571428571428573,
		18.181818181818183
	],
	"4": [
		37.5,
		22.22222222222222,
		0,
		0,
		16.666666666666668,
		0,
		0,
		14.285714285714286,
		16.666666666666668,
		25,
		0,
		15.789473684210526,
		28.571428571428573,
		0,
		14.285714285714286,
		9.090909090909092
	],
	"5": [
		0,
		11.11111111111111,
		25,
		28.571428571428573,
		33.333333333333336,
		42.857142857142854,
		50,
		0,
		0,
		25,
		9.090909090909092,
		0,
		28.571428571428573,
		42.857142857142854,
		14.285714285714286,
		63.63636363636363
	],
	"6": [
		12.5,
		11.11111111111111,
		25,
		28.571428571428573,
		33.333333333333336,
		0,
		0,
		28.571428571428573,
		33.333333333333336,
		25,
		27.272727272727273,
		21.05263157894737,
		28.571428571428573,
		14.285714285714286,
		0,
		9.090909090909092
	],
	"7": [
		25,
		22.22222222222222,
		0,
		0,
		0,
		0,
		33.333333333333336,
		14.285714285714286,
		16.666666666666668,
		12.5,
		0,
		5.2631578947368425,
		0,
		14.285714285714286,
		14.285714285714286,
		0
	],
	"No loss": [
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0,
		0
	]
} as const;