// generated-by-analysis/Pilot10/graph-inputs/harm-quantity-with-likert-severity-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/harm-quantity-with-likert-severity.ts.
// spell-checker: disable

	export const LostPhotos = {
	"labels": [
		"< 1 Day",
		"Days",
		"Weeks",
		"Months",
		"1-5 Years",
		"> 5 Years"
	],
	"counts": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			1,
			0,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0
		]
	},
	"percentsOfResponses": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			11.11111111111111,
			0,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0
		]
	}
} as const;
	export const LostEmails = {
	"labels": [
		"< 1 Day",
		"Days",
		"Weeks",
		"Months",
		"1-5 Years",
		"> 5 Years"
	],
	"counts": {
		"1": [
			1,
			0,
			0,
			0,
			0,
			1
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			1
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0
		]
	},
	"percentsOfResponses": {
		"1": [
			11.11111111111111,
			0,
			0,
			0,
			0,
			11.11111111111111
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			11.11111111111111
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0
		]
	}
} as const;
	export const MentalHealth = {
	"labels": [
		"< 1 Day",
		"Days",
		"Weeks",
		"Months",
		"1-5 Years",
		"> 5 Years"
	],
	"counts": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0
		]
	},
	"percentsOfResponses": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"3": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0
		]
	}
} as const;