// generated-by-analysis/Pilot10/graph-inputs/harm-focused-likert-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/harm-focused-likert.ts.
// spell-checker: disable

	export const labels = [
	"Lost Photos",
	"Lost Emails",
	"Mental Health"
] as const;
	export const counts = {
	"1": [
		0,
		2,
		0
	],
	"2": [
		0,
		0,
		0
	],
	"3": [
		0,
		1,
		1
	],
	"4": [
		0,
		0,
		0
	],
	"5": [
		1,
		0,
		1
	],
	"6": [
		0,
		0,
		1
	],
	"7": [
		0,
		0,
		0
	],
	"No loss": [
		0,
		0,
		0
	]
} as const;
	export const percentsOfResponses = {
	"1": [
		0,
		22.22222222222222,
		0
	],
	"2": [
		0,
		0,
		0
	],
	"3": [
		0,
		11.11111111111111,
		11.11111111111111
	],
	"4": [
		0,
		0,
		0
	],
	"5": [
		11.11111111111111,
		0,
		11.11111111111111
	],
	"6": [
		0,
		0,
		11.11111111111111
	],
	"7": [
		0,
		0,
		0
	],
	"No loss": [
		0,
		0,
		0
	]
} as const;
	export const percentsOfAnswered = {
	"1": [
		0,
		66.66666666666667,
		0
	],
	"2": [
		0,
		0,
		0
	],
	"3": [
		0,
		33.333333333333336,
		33.333333333333336
	],
	"4": [
		0,
		0,
		0
	],
	"5": [
		100,
		0,
		33.333333333333336
	],
	"6": [
		0,
		0,
		33.333333333333336
	],
	"7": [
		0,
		0,
		0
	],
	"No loss": [
		0,
		0,
		0
	]
} as const;