// generated-by-analysis/Pilot10/graph-inputs/demographics-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/demographics.ts.
// spell-checker: disable

export const gender = {
	"labels": [
		"Male",
		"Female",
		"Non-binary/third",
		"Self-describe",
		"Prefer not to say"
	],
	"tallies": {
		"Male": 13,
		"_totalAnswered": 25,
		"Female": 11,
		"Non-binary/third": 1
	},
	"data": [
		13,
		11,
		1,
		0,
		0
	]
} as const;
export const genderPercent = {
	"labels": [
		"Male",
		"Female",
		"Non-binary/third",
		"Self-describe",
		"Prefer not to say"
	],
	"tallies": {
		"Male": 52,
		"Female": 44,
		"Non-binary/third": 4
	},
	"data": [
		52,
		44,
		4,
		0,
		0
	]
} as const;
export const education = {
	"labels": [
		"< High School",
		"High School",
		"Some College",
		"Associate",
		"Bachelor's",
		"Master's",
		"Professional",
		"Doctoral",
		"Decline to Answer"
	],
	"tallies": {
		"High School": 3,
		"_totalAnswered": 25,
		"Doctoral": 1,
		"< High School": 1,
		"Bachelor's": 9,
		"Some College": 4,
		"Master's": 2,
		"Associate": 4,
		"Professional": 1
	},
	"data": [
		1,
		3,
		4,
		4,
		9,
		2,
		1,
		1,
		0
	]
} as const;
export const educationPercent = {
	"labels": [
		"< High School",
		"High School",
		"Some College",
		"Associate",
		"Bachelor's",
		"Master's",
		"Professional",
		"Doctoral",
		"Decline to Answer"
	],
	"tallies": {
		"High School": 12,
		"Doctoral": 4,
		"< High School": 4,
		"Bachelor's": 36,
		"Some College": 16,
		"Master's": 8,
		"Associate": 16,
		"Professional": 4
	},
	"data": [
		4,
		12,
		16,
		16,
		36,
		8,
		4,
		4,
		0
	]
} as const;